MS-5 Crear modulo de usuarios - wip

+++ Modules/User/Http/Controllers/UserController.php
@@ -7,8 +7,9 @@ use App\Http\Requests\UpdateUser;
 use App\User;
 use Illuminate\Http\Request;
 use Illuminate\Http\Response;
-use Illuminate\Routing\Controller;
+//use Illuminate\Routing\Controller;
 use Styde\Html\Facades\Alert;
+use App\Http\Controllers\Controller;
 
 class UserController extends Controller
 {
@@ -18,6 +19,8 @@ class UserController extends Controller
      */
     public function index()
     {
+        $this->authorize('view', User::class);
+
         $users = User::all();
         return view('user::index', compact('users'));
     }
@@ -28,6 +31,8 @@ class UserController extends Controller
      */
     public function create()
     {
+        $this->authorize('create', User::class);
+
         return view('user::create');
     }
 
@@ -38,6 +43,8 @@ class UserController extends Controller
      */
     public function store(StoreUser $request)
     {
+        $this->authorize('create', User::class);
+
         $newUser = new User();
         $newUser->name = $request->input('Nombre');
         $newUser->email = $request->input('Email');
@@ -66,6 +73,7 @@ class UserController extends Controller
      */
     public function edit($id)
     {
+        $this->authorize('update', User::class);
         $user = User::findOrFail($id);
         return view('user::edit',compact('user'));
     }
@@ -75,8 +83,34 @@ class UserController extends Controller
      * @param  Request $request
      * @return Response
      */
-    public function update(UpdateUser $request)
+    public function update(Request $request)
     {
+        $this->authorize('update', User::class);
+
+        $user = User::findOrFail($request->input('id'));
+        $name = $request->input('Nombre');
+        $email = $request->input('Email');
+        $password = $request->input('password');
+
+        $request->validate([
+            'Nombre' => 'required',
+            'Email' => 'required|unique:users,email,'.$user->id,
+            'oldPassword' => 'required_with:password',
+            'password' => 'required_with:oldPassword',
+            'password2' => 'required_with:password|same:password',
+        ]);
+
+        $user->name = $name;
+        $user->email = $email;
+        if (!is_null($password)) {
+            $user->password = \Hash::make($password);
+        }
+        $user->save();
+
+        Alert::success('El usuario fue modificado correctamente.');
+
+        return redirect()->route('user.index');
+
     }
 
     /**
@@ -85,12 +119,18 @@ class UserController extends Controller
      */
     public function destroy(Request $request ,$id)
     {
+        $this->authorize('delete', User::class);
+
         $user = User::findOrFail($id);
         $user->delete();
 
-        Alert::success('El usuario fue borrado correctamente.');
-
-        return redirect()->route('user.index');
+        $message = 'El usuario fue borrado correctamente.';
+        if ($request->ajax()) {
+            return response()->json(['success'=>true, 'data'=>['message'=>$message]]);
+        }else{
+            Alert::success($message);
+            return redirect()->route('user.index');
+        }
 
     }
 }
